Port all to new TitleManager, fix inventory location
